Add a font-size mixin and px to rem function
Set 16px as the base font size.

Add a function to convert pixels to rem.

Use the rem function in the font-size mixin, so a pixel value can be
passed in and a pixel fallback will be output with rem.

For now, create core- mixins for the font sizes we are using. This will
give us an idea which of the core typography styles are used by
components.