Add old IE helpers
These are required by the button component as it uses the ie-lte syntax
for IE versions lower than or equal to 8.

Set is-ie to true to output IE specific styles in the govuk-frontend
stylesheet.

We will need to decide how best to support earlier versions of IE - are
the overrides necessary for ‘functional’ support?