Allow for text argument to be passed by itself
This is a bit of syntactic sugar to make it easier to use error messages where the user does not want to specify classes or extra attributes nor use HTML.

It allows the user to call e.g. `govukErrorMessage("Full name must be provided")` without having to pass an entire object.

We do this in the macro so that if in the future we transpile the templates we do not have to worry about mimicking this functionality in other languages.